requirement.txt
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
numpy==1.24.3
pillow==9.5.0
matplotlib==3.7.1
opencv-python==4.7.0.72
scikit-image==0.20.0
scipy==1.10.1
tqdm==4.65.0
albumentations==1.3.1

#-e .